off_host_apps/TrafficLightActuator/requirements.txt
paho-mqtt==2.1.0
RPi.GPIO==0.7.1